- added:     - logger in repository;     - js for update table and add, edit and delete car;     - new method in controller for statistic base.

+++ src/main/java/dmitrish/com/github/directoryOfCar/controller/rest/RestControllerCar.java
@@ -20,10 +20,14 @@ public class RestControllerCar {
         this.carService = carService;
     }
 
+
     @PostMapping(value = "/add")
     public ResponseEntity<?> addCar(@RequestBody Car car) {
-        carService.add(car);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (!carService.isExist(car)) {
+            carService.add(car);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(car.getLicensePlate() + " already exist in base", HttpStatus.BAD_REQUEST);
     }
 
     @GetMapping(value = "/list")
@@ -37,21 +41,56 @@ public class RestControllerCar {
     @GetMapping("/{id}")
     public ResponseEntity<Car> getCarById(@PathVariable(name = "id") long id) {
         final Car car = carService.getById(id);
-
         return car != null
                 ? new ResponseEntity<>(car, HttpStatus.OK)
                 : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @PutMapping("/update")
-    public ResponseEntity<?> updateCar (@RequestBody Car car) {
-        carService.update(car);
+    public ResponseEntity<?> updateCar(@RequestBody Car car) {
+        Car fromBase = carService.getById(car.getId());
+        fromBase.setLicensePlate(car.getLicensePlate());
+        fromBase.setBrand(car.getBrand());
+        fromBase.setModel(car.getModel());
+        fromBase.setColor(car.getColor());
+        fromBase.setPrice(car.getPrice());
+        fromBase.setYearOfManufacture(car.getYearOfManufacture());
+        carService.update(fromBase);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<?> deleteCar(@RequestBody Car car) {
-        carService.delete(car);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<?> deleteCar(@RequestBody Long id) {
+        Car car = carService.getById(id);
+        if (car == null) {
+            return new ResponseEntity<>("Car id: " + id + ", not found in base", HttpStatus.BAD_REQUEST);
+        } else {
+            carService.delete(car);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+    }
+
+    @GetMapping("/stat/notes")
+    public ResponseEntity<Integer> getCountOfNotes() {
+        int count = carService.countOfNotes();
+        return (count != 0)
+                ? new ResponseEntity<>(count, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    @GetMapping("/stat/notes/older/{year}")
+    public ResponseEntity<Integer> getCountOfCarOlderYear(@PathVariable(value = "year") Integer year) {
+        int count = carService.countOfCarOlderYear(year);
+        return (count != 0)
+                ? new ResponseEntity<>(count, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    @GetMapping("/stat/notes/until/{year}")
+    public ResponseEntity<Integer> getCountOfCarUntilYear(@PathVariable(value = "year") Integer year) {
+        int count = carService.countOfCarUntilYear(year);
+        return (count != 0)
+                ? new ResponseEntity<>(count, HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
